+++ shoreline.go
@@ -43,9 +43,14 @@ func main() {
 	}
 
 	// server secret may be passed via a separate env variable to accomodate easy secrets injection via Kubernetes
+	// The server secret is the password any Tidepool service is supposed to know and pass to shoreline for authentication and for getting token
+	// With Mdblp, we consider we can have different server secrets
+	// These secrets are hosted in a map[string][string] instead of single string
+	// which 1st string represents Server/Service name and 2nd represents the actual secret
+	// here we consider this SERVER_SECRET that can be injected via Kubernetes is the one for the default server/service (any Tidepool service)
 	serverSecret, found := os.LookupEnv("SERVER_SECRET")
 	if found {
-		config.User.ServerSecret = serverSecret
+		config.User.ServerSecrets["default"] = serverSecret
 	}
 
 	userSecret, found := os.LookupEnv("API_SECRET")
@@ -137,7 +142,7 @@ func main() {
 	 * User-Api setup
 	 */
 
-	log.Print(shoreline_service_prefix, "adding", user.USER_API_PREFIX)
+	log.Print(shoreline_service_prefix, "adding ", user.USER_API_PREFIX)
 
 	userapi := user.InitApi(config.User, user.NewMongoStoreClient(&config.Mongo), highwater)
 	userapi.SetHandlers("", rtr)
@@ -150,14 +155,14 @@ func main() {
 		WithTokenProvider(userClient).
 		Build()
 
-	log.Print(shoreline_service_prefix, "adding", "permsClient")
+	log.Print(shoreline_service_prefix, "adding ", "permsClient")
 	userapi.AttachPerms(permsClient)
 
 	/*
 	 * Oauth setup
 	 */
 
-	log.Print(shoreline_service_prefix, "adding", oauth2.OAUTH2_API_PREFIX)
+	log.Print(shoreline_service_prefix, "adding ", oauth2.OAUTH2_API_PREFIX)
 
 	oauthapi := oauth2.InitApi(config.Oauth2, oauth2.NewOAuthStorage(&config.Mongo), userClient, permsClient)
 	oauthapi.SetHandlers("", rtr)