+++ user/hasher_test.go
@@ -22,7 +22,7 @@ func TestGenerateUniqueHashLength(t *testing.T) {
 
 	if theHash, err := generateUniqueHash([]string{"one", "two", "miss a few", "99", "100"}, 20); err == nil {
 		if len(theHash) != 20 {
-			t.Fatalf("the has should be 20 characters in length, but has %v", len(theHash))
+			t.Fatalf("the hash should be 20 characters in length, but has %v", len(theHash))
 		}
 	} else {
 		t.Fatalf("there should be no error given")